# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Belconnen United FC Performance Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /matches:
    get:
      operationId: listMatches
      tags: [matches]
      summary: List matches
      parameters:
        - name: teamId
          in: query
          schema: { type: integer }
        - name: seasonId
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of matches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
    post:
      operationId: createMatch
      tags: [matches]
      summary: Create a match record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchInput"
      responses:
        "201":
          description: Created match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
  /matches/{id}:
    get:
      operationId: getMatch
      tags: [matches]
      summary: Get a match
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
    patch:
      operationId: updateMatch
      tags: [matches]
      summary: Update a match
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchUpdate"
      responses:
        "200":
          description: Updated match
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"

  # ─── PLAYER MATCH STATS ─────────────────────────────────────────────────────
  /player-stats:
    get:
      operationId: listPlayerStats
      tags: [player-stats]
      summary: List player match stats
      parameters:
        - name: matchId
          in: query
          schema: { type: integer }
        - name: playerId
          in: query
          schema: { type: integer }
        - name: teamId
          in: query
          schema: { type: integer }
        - name: seasonId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of player stats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerStat"
    post:
      operationId: createPlayerStat
      tags: [player-stats]
      summary: Record player stats for a match
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerStatInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlayerStat"

  # ─── GOALS ──────────────────────────────────────────────────────────────────
  /goals:
    get:
      operationId: listGoals
      tags: [goals]
      summary: List goal events
      parameters:
        - name: matchId
          in: query
          schema: { type: integer }
        - name: teamId
          in: query
          schema: { type: integer }
        - name: seasonId
          in: query
          schema: { type: integer }
        - name: scorerTeam
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of goal events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Goal"
    post:
      operationId: createGoal
      tags: [goals]
      summary: Record a goal event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
  /goals/{id}:
    patch:
      operationId: updateGoal
      tags: [goals]
      summary: Update a goal event (e.g. set X/Y coordinates)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GoalUpdate"
      responses:
        "200":
          description: Updated goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Goal"
    delete:
      operationId: deleteGoal
      tags: [goals]
      summary: Delete a goal event
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ─── GPS SESSIONS ───────────────────────────────────────────────────────────
  /gps-sessions:
    get:
      operationId: listGpsSessions
      tags: [gps]
      summary: List GPS sessions
      parameters:
        - name: playerId
          in: query
          schema: { type: integer }
        - name: year
          in: query
          schema: { type: string }
        - name: teamId
          in: query
          schema: { type: integer }
        - name: round
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of GPS sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GpsSession"
    post:
      operationId: createGpsSession
      tags: [gps]
      summary: Create a GPS session record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GpsSessionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GpsSession"

  # ─── ATHLETIC TESTING ───────────────────────────────────────────────────────
  /athletic-tests:
    get:
      operationId: listAthleticTests
      tags: [testing]
      summary: List athletic test results
      parameters:
        - name: playerId
          in: query
          schema: { type: integer }
        - name: year
          in: query
          schema: { type: string }
        - name: teamId
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: List of athletic test results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AthleticTest"
    post:
      operationId: createAthleticTest
      tags: [testing]
      summary: Record an athletic test result
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AthleticTestInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AthleticTest"

  # ─── ANALYTICS (WOW endpoints) ──────────────────────────────────────────────
...
components:
  schemas:
...
    Match:
      type: object
      properties:
        id:             { type: integer }
        matchId:        { type: string }
        matchDate:      { type: ["string", "null"] }
        venue:          { type: ["string", "null"] }
        opponent:       { type: string }
        halfScore:      { type: ["string", "null"] }
        fullScore:      { type: ["string", "null"] }
        goalsScored:    { type: ["integer", "null"] }
        goalsConceded:  { type: ["integer", "null"] }
        cleanSheet:     { type: ["boolean", "null"] }
        formation:      { type: ["string", "null"] }
        oppFormation:   { type: ["string", "null"] }
        conditions:     { type: ["string", "null"] }
        possession:     { type: ["number", "null"] }
        shots:          { type: ["integer", "null"] }
        passes:         { type: ["integer", "null"] }
        oppShots:       { type: ["integer", "null"] }
        oppPasses:      { type: ["integer", "null"] }
        quadrantPoints: { type: ["string", "null"] }
        teamId:         { type: integer }
        seasonId:       { type: integer }
        result:         { type: ["string", "null"] }
      required: [id, matchId, opponent, teamId, seasonId]
    MatchInput:
      type: object
      properties:
        matchId:        { type: string }
        matchDate:      { type: ["string", "null"] }
        venue:          { type: ["string", "null"] }
        opponent:       { type: string }
        halfScore:      { type: ["string", "null"] }
        fullScore:      { type: ["string", "null"] }
        goalsScored:    { type: ["integer", "null"] }
        goalsConceded:  { type: ["integer", "null"] }
        cleanSheet:     { type: ["boolean", "null"] }
        formation:      { type: ["string", "null"] }
        oppFormation:   { type: ["string", "null"] }
        conditions:     { type: ["string", "null"] }
        possession:     { type: ["number", "null"] }
        shots:          { type: ["integer", "null"] }
        passes:         { type: ["integer", "null"] }
        oppShots:       { type: ["integer", "null"] }
        oppPasses:      { type: ["integer", "null"] }
        quadrantPoints: { type: ["string", "null"] }
        teamId:         { type: integer }
        seasonId:       { type: integer }
      required: [matchId, opponent, teamId, seasonId]
    MatchUpdate:
      type: object
      properties:
        matchDate:      { type: ["string", "null"] }
        venue:          { type: ["string", "null"] }
        opponent:       { type: string }
        halfScore:      { type: ["string", "null"] }
        fullScore:      { type: ["string", "null"] }
        goalsScored:    { type: ["integer", "null"] }
        goalsConceded:  { type: ["integer", "null"] }
        cleanSheet:     { type: ["boolean", "null"] }
        formation:      { type: ["string", "null"] }
        oppFormation:   { type: ["string", "null"] }
        conditions:     { type: ["string", "null"] }
        possession:     { type: ["number", "null"] }
        shots:          { type: ["integer", "null"] }
        passes:         { type: ["integer", "null"] }
        oppShots:       { type: ["integer", "null"] }
        oppPasses:      { type: ["integer", "null"] }
        quadrantPoints: { type: ["string", "null"] }

    # ── Player Stats ──────────────────────────────────────────────────────────
    PlayerStat:
      type: object
      properties:
        id:          { type: integer }
        matchId:     { type: integer }
        playerId:    { type: integer }
        playerName:  { type: string }
        minsPlayed:  { type: ["integer", "null"] }
        position:    { type: ["string", "null"] }
        discipline:  { type: ["string", "null"] }
        started:     { type: ["boolean", "null"] }
        appearance:  { type: ["boolean", "null"] }
        club:        { type: ["string", "null"] }
        year:        { type: ["string", "null"] }
      required: [id, matchId, playerId, playerName]
    PlayerStatInput:
      type: object
      properties:
        matchId:    { type: integer }
        playerId:   { type: integer }
        minsPlayed: { type: ["integer", "null"] }
        position:   { type: ["string", "null"] }
        discipline: { type: ["string", "null"] }
        started:    { type: ["boolean", "null"] }
        appearance: { type: ["boolean", "null"] }
        club:       { type: ["string", "null"] }
        year:       { type: ["string", "null"] }
      required: [matchId, playerId]

    # ── Goals ─────────────────────────────────────────────────────────────────
    Goal:
      type: object
      properties:
        id:               { type: integer }
        matchId:          { type: integer }
        recording:        { type: ["string", "null"] }
        matchDate:        { type: ["string", "null"] }
        homeTeam:         { type: ["string", "null"] }
        awayTeam:         { type: ["string", "null"] }
        scorerTeam:       { type: ["string", "null"] }
        minuteScored:     { type: ["integer", "null"] }
        scorer:           { type: ["string", "null"] }
        assist:           { type: ["string", "null"] }
        goalType:         { type: ["string", "null"] }
        assistType:       { type: ["string", "null"] }
        howPenetrated:    { type: ["string", "null"] }
        buildupLane:      { type: ["string", "null"] }
        firstTimeFinish:  { type: ["boolean", "null"] }
        finishType:       { type: ["string", "null"] }
        passString:       { type: ["string", "null"] }
        goalX:            { type: ["number", "null"] }
        goalY:            { type: ["number", "null"] }
        teamId:           { type: integer }
        seasonId:         { type: integer }
      required: [id, matchId, teamId, seasonId]
    GoalInput:
      type: object
      properties:
        matchId:          { type: integer }
        recording:        { type: ["string", "null"] }
        matchDate:        { type: ["string", "null"] }
        homeTeam:         { type: ["string", "null"] }
        awayTeam:         { type: ["string", "null"] }
        scorerTeam:       { type: ["string", "null"] }
        minuteScored:     { type: ["integer", "null"] }
        scorer:           { type: ["string", "null"] }
        assist:           { type: ["string", "null"] }
        goalType:         { type: ["string", "null"] }
        assistType:       { type: ["string", "null"] }
        howPenetrated:    { type: ["string", "null"] }
        buildupLane:      { type: ["string", "null"] }
        firstTimeFinish:  { type: ["boolean", "null"] }
        finishType:       { type: ["string", "null"] }
        passString:       { type: ["string", "null"] }
        goalX:            { type: ["number", "null"] }
        goalY:            { type: ["number", "null"] }
        teamId:           { type: integer }
        seasonId:         { type: integer }
      required: [matchId, teamId, seasonId]
    GoalUpdate:
      type: object
      properties:
        minuteScored:    { type: ["integer", "null"] }
        scorer:          { type: ["string", "null"] }
        assist:          { type: ["string", "null"] }
        goalType:        { type: ["string", "null"] }
        assistType:      { type: ["string", "null"] }
        howPenetrated:   { type: ["string", "null"] }
        buildupLane:     { type: ["string", "null"] }
        firstTimeFinish: { type: ["boolean", "null"] }
        finishType:      { type: ["string", "null"] }
        passString:      { type: ["string", "null"] }
        goalX:           { type: ["number", "null"] }
        goalY:           { type: ["number", "null"] }

    # ── GPS Sessions ──────────────────────────────────────────────────────────
    GpsSession:
      type: object
      properties:
        id:                  { type: integer }
        sessionDate:         { type: ["string", "null"] }
        sessionTitle:        { type: ["string", "null"] }
        playerName:          { type: string }
        playerId:            { type: ["integer", "null"] }
        teamId:              { type: integer }
        year:                { type: string }
        round:               { type: ["string", "null"] }
        opponent:            { type: ["string", "null"] }
        splitName:           { type: ["string", "null"] }
        tags:                { type: ["string", "null"] }
        minsPlayed:          { type: ["number", "null"] }
        distanceKm:          { type: ["number", "null"] }
        sprintDistanceM:     { type: ["number", "null"] }
        powerPlays:          { type: ["number", "null"] }
        energyKcal:          { type: ["number", "null"] }
        impacts:             { type: ["number", "null"] }
        hrLoad:              { type: ["number", "null"] }
        timeInRedZoneMin:    { type: ["number", "null"] }
        playerLoad:          { type: ["number", "null"] }
        topSpeedMs:          { type: ["number", "null"] }
        distancePerMinMm:    { type: ["number", "null"] }
        powerScoreWkg:       { type: ["number", "null"] }
        workRatio:           { type: ["number", "null"] }
        hrMaxBpm:            { type: ["number", "null"] }
        maxDecelerationMss:  { type: ["number", "null"] }
        maxAccelerationMss:  { type: ["number", "null"] }
        distanceZone1Km:     { type: ["number", "null"] }
        distanceZone2Km:     { type: ["number", "null"] }
        distanceZone3Km:     { type: ["number", "null"] }
        distanceZone4Km:     { type: ["number", "null"] }
        distanceZone5Km:     { type: ["number", "null"] }
      required: [id, playerName, teamId, year]
    GpsSessionInput:
      type: object
      properties:
        sessionDate:         { type: ["string", "null"] }
        sessionTitle:        { type: ["string", "null"] }
        playerName:          { type: string }
        playerId:            { type: ["integer", "null"] }
        teamId:              { type: integer }
        year:                { type: string }
        round:               { type: ["string", "null"] }
        opponent:            { type: ["string", "null"] }
        splitName:           { type: ["string", "null"] }
        tags:                { type: ["string", "null"] }
        minsPlayed:          { type: ["number", "null"] }
        distanceKm:          { type: ["number", "null"] }
        sprintDistanceM:     { type: ["number", "null"] }
        powerPlays:          { type: ["number", "null"] }
        energyKcal:          { type: ["number", "null"] }
        impacts:             { type: ["number", "null"] }
        hrLoad:              { type: ["number", "null"] }
        timeInRedZoneMin:    { type: ["number", "null"] }
        playerLoad:          { type: ["number", "null"] }
        topSpeedMs:          { type: ["number", "null"] }
        distancePerMinMm:    { type: ["number", "null"] }
        powerScoreWkg:       { type: ["number", "null"] }
        workRatio:           { type: ["number", "null"] }
        hrMaxBpm:            { type: ["number", "null"] }
        maxDecelerationMss:  { type: ["number", "null"] }
        maxAccelerationMss:  { type: ["number", "null"] }
        distanceZone1Km:     { type: ["number", "null"] }
        distanceZone2Km:     { type: ["number", "null"] }
        distanceZone3Km:     { type: ["number", "null"] }
        distanceZone4Km:     { type: ["number", "null"] }
        distanceZone5Km:     { type: ["number", "null"] }
      required: [playerName, teamId, year]

    # ── Athletic Tests ────────────────────────────────────────────────────────
...