# lib/api-zod/src/generated/types/playerStat.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Belconnen United FC Performance Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerStat {
  id: number;
  matchId: number;
  playerId: number;
  playerName: string;
  /** @nullable */
  minsPlayed?: number | null;
  /** @nullable */
  position?: string | null;
  /** @nullable */
  discipline?: string | null;
  /** @nullable */
  started?: boolean | null;
  /** @nullable */
  appearance?: boolean | null;
  /** @nullable */
  club?: string | null;
  /** @nullable */
  year?: string | null;
}